Refactor AntPolarsLoop to improve readability by consolidating method chaining in get_best_moves

+++ examples/sugarscape_ig/ss_polars/agents.py
@@ -258,13 +258,12 @@ class AntPolarsLoop(AntPolarsBase):
         best_moves = (
             # Only fill nulls for the column we need as an int sentinel;
             # avoid touching UInt columns like 'blocking_agent_id'.
-            neighborhood.with_columns(
-                pl.col("blocking_agent_order").fill_null(-1)
-            )
+            neighborhood.with_columns(pl.col("blocking_agent_order").fill_null(-1))
             .cast({"agent_order": pl.Int32, "blocking_agent_order": pl.Int32})
             .select(
                 pl.struct(["agent_order", "blocking_agent_order"]).map_batches(
-                    map_batches_func
+                    map_batches_func,
+                    return_dtype=pl.Int32,
                 )
             )
             .with_columns(
